Refactor: move factories to factory subpackage
